requirements.txt
requests
python-dotenv
authlib
deezer-python
